Add files via upload
Databases are being used in every aspect of our lives right now. Trillions of bytes of data are being stored in servers around the world. SQL is one of the most basic methods to use such a database. But have you ever thought about using C++ to maintain such a database. In this post, we will talk about implementing different views on a text file according to the type of user and edit accordingly.

The data stored using this code are: 
1) Registration number 
2) Name 
3) Marks in CSE1001 
4) Marks in CSE1002 
5) Proctor ID